alternator: apply default_time_to_live to update operations
CQL tables have a default_time_to_live parameter which allows
specifying a default TTL, which gets automatically applied
for each insert/update. Alternator used to ignore that parameter,
but it might be useful to set it via CQL and have an alternator
table which automatically applies time to live to all updates.