install-dependencies.sh: rename pip_packages & pip_symlinks lists
The pip_packages and pip_symlinks lists are confusing as they have
nothing with the packages installed by this script but are used by the
scylla-python3 project to install packages on scylla's custom python-env.

This commit rename both lists to make it clear about their usage

The reason for this change is the need to maintain a list of python
packages (aka pip) that will be installed by this script - will be added
later on another commit.