install-dependencies.sh: install python packages by pip instead of dnf
The install-dependencies script install python script by dnf which in
some situations - ex, pytest_asyncio - blocks us from installing a newer
version than what is in the dnf repo,

This commit moves a few python packages to be installed by pip. There are
a few more that can be changed, but let's start with these and update the
rest later.

Refs #11155 #10550