perf_simple_query: add attach/detach points for perf
Instrumenting `perf_simple_query` with perf is tricky as the startup and
teardown code introduces a lot of unpredictable and undesired noise into
measurements. One can try to work around this by attaching perf at the
runtime, after the measure cycle has actually started, and cancel right
before or right after it ends. But even this is unreliable when one
wants to do a perf-diff between two runs as in this case the two runs
should be as identical as possible and even the time at which the
attach/detach happens can introduce noise.
This patch offers a crude yet effective solution: when the new
--perf-attach-detach-point flag is used, there will be a 5 second wait
just before/after starting/ending the measure cycle. During this wait
time the system is effectively idle. This time should be enough to
quickly alt-tab into another terminal and run a prepared perf-attach
statement and to CTRL+C at the end.