database: truncate: fix assert about replay_position low_mark
This assert was tweaked several times:
Introduced in 83323e155e2f26ead355b45b64531980e8ca66ff,
then fixed in b2b1a1f7e1b75afb135b3bcacd4be33ce10d57c4 to account
for no rp from discard_sstables, then in
9620755c7f2b18109dc4b2b64a0cd38ff0cbb2eb to account for
cases we do not flush the table, then again in
71c5dc82df7eb7f63c1be4ef07d569bf0e97b5e5 to make that more accurate.

But, the assert wasn't correct in the first place
in the sense that we first get `low_mark` which
represents the highest replay_position at the time truncate
was called, but then we call discard_sstables with a time_point
of `truncated_at` that we get from the caller via the timestamp_func,
and that one could be in the past, before truncate was called -
hence discard_sstables with that timestamp may very well
return a replay_position from older sstables, prior to flush
that can be smaller than the low_mark.

Fix this assert to account for that case.

The real fix to this issue is to have a truncate_tombstone
that will carry an authoritative api::timstamp (#11230)

Fixes #11231

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>