service: storage_proxy: don't access `remote` when calculating target replicas for local queries
We only want to access `remote` when it's necessary - when we're
performing a query that involves remote nodes. We want to support local
queries when `remote` (in particular, `gossiper&`) is unavailable.

Add a helper, `calculate_target_replicas`, which will check if it's a
local query and return early in that case without accessing `remote`.