utils/phased_barrier: add close()
Behaving exactly like gate::close(): in fact all this method does is
close the current gate and not replace it with a new one, such that
later attempts to call `start()` or `advance_and_await()` will result
in `gate_closed` exception. This allows a phaser to be closed at the
end of its lifecycle, disallowing new future operations.