replica/table: close_for_read_write_ops()
The current `await_pending_ops()` has a weakness: it doesn't prevent new
ops from being started after it is called. So it is possible for new
reads/writes/streams to be started even after a table is detached from
the database, if the user has a table reference at hand.
This patch fixes this: `await_pending_ops()` is renamed to
`close_for_read_write_ops()` and it uses the new phased_barrier::close()
to close the read/write/stream phasers, preventing any such operations
from being started. We don't want to do the same with the flush phaser
as truncated (done as part of table drop) wants to start new flushes
after the table is detached. So we remove this phaser from said method
and users have to wait for flushes explicitly for now.