replica/table: replace _async_gate with _generic_op_phaser
At this point only table-internal usages remain. Migrate all these to
the generic phaser. This prepares the ground for dropping the async
gate. having all phasers will make it more clear that one has to use
just one of them, not a phaser *and* the gate, which is what previously
seem to have been the case (but wasn't respected uniformly).
Up to this point `_async_gate.is_closed()` indicated whether the table
was stopped. Its replacement, `_pending_generic_ops_phaser.is_closed()`
cannot serve the same purpose, as it is not the first thing we stop in
`table::stop()` (unlike async gate). To bridge this, we add an explicit
`_stopped` flag.