Adding the __level labels
Scylla generates many metrics, and when multiplied by the number of
shards, the total number of metrics adds a significant load to a
monitoring server.

With multi-tier monitoring, it is helpful to have a smaller subset of
metrics users care about and allow them to get only those.

This patch adds a __level label, currently with three values, but we can
add more in the future.  We will use the label to mark the
interesting user-facing metrics.

The three levels are:
basic - metrics for Scylla monitoring
advanced - metrics for advanced debugging situations
specific - metrics for specific high cardinality metrics, like per table
metrics.

In this phase, basic will mark all metrics used in the dashboards.
specific,  would mark the per-table metrics that are disabled by default.
advanced is here for future use.
In practice, without any configuration change, Prometheus would get the
same metrics as it gets today.

While it is possible to filter by the label, e.g.:
curl http://localhost:9180/metrics?__level=1

The labels themselves are not reported thanks to label filtering of
labels begin with __.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>