Revert "distributed_loader: Remove unused load-prio manipulations"
This reverts commit 7396de72b17ab15154f5b994e5fc3851a033d6be.

In 7396de7 (and refactorings before it) the set of prioritized keyspaces (and processing thereof)
was removed, due to apparent non-usage (which is true for open-source version).

This functionality is however required for certain features of the enterprise version (ear).
As such is needs to be restored and reenabled. This reverts the actual commit, patch after
ensures we use the prio set.